pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Augusto Balsanelli", email = "augustebalsanel@proton.me"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "setuptools>=68.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
